requirements.txt
streamlit==1.56.0
transformers==5.6.2
torch==2.11.0
sumy==0.12.0
nltk==3.9.4
rouge-score==0.1.2
openai-whisper==20250625
youtube-transcript-api==1.2.4
pypdf==6.4.0
pydub==0.25.1
numpy==2.4.4
pytest==9.0.2
